requirements.txt
boto3==1.17.71
botocore==1.20.71
et-xmlfile==1.1.0
jmespath==0.10.0
openpyxl==3.0.7
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
